#55 Updated the average speed computing functionality for each Ski Run
