+++ src/main/java/de/dennisguse/opentracks/stats/TrackStatistics.java
@@ -56,6 +56,7 @@
      private Duration movingTime;
      // The maximum speed (meters/second) that we believe is valid.
      private Speed maxSpeed;
+     private double averageSpeed;
      private Float totalAltitudeGain_m = null;
      private Float totalAltitudeLoss_m = null;
 
@@ -393,6 +394,15 @@
          this.maximumSpeedPerRun = maximumSpeedPerRun;
      }
 
+     public double getAverageSpeedPerRun() {
+         return averageSpeed;
+     }
+
+     public void setAverageSpeedPerRun(double speed) {
+         this.averageSpeed = speed;
+
+     }
+
      @VisibleForTesting(otherwise = VisibleForTesting.PACKAGE_PRIVATE)
      public void addTotalAltitudeLoss(float loss_m) {
          if (totalAltitudeLoss_m == null) {
@@ -415,6 +425,7 @@
          return "TrackStatistics { Start Time: " + getStartTime() + "; Stop Time: " + getStopTime()
                  + "; Total Distance: " + getTotalDistance() + "; Total Time: " + getTotalTime()
                  + "; Moving Time: " + getMovingTime() + "; Max Speed: " + getMaxSpeed()
+                 + "; Average Speed Per Run: " + getAverageSpeedPerRun()
                  + "; Maximum Speed Per Run: " + getMaximumSpeedPerRun()
                  + "; Min Altitude: " + getMinAltitude() + "; Max Altitude: " + getMaxAltitude()
                  + "; Altitude Gain: " + getTotalAltitudeGain()
